Refactor core of visu
This refactor of the visu.core module aims at providing more flexibility
for the creation of visualizations with complex layouts.

The approach relying on the exposition of the Axes does not allow
nesting layouts.  Rather, the approach introduced by this commit exposes
both the figure and a SubplotSpec.  As a result, a visualization is able
to freely manage its allocated subspace.